Move RedirectFixer to redirect package. New ResolverInfos class
