+++ app/src/main/java/com/tasomaniac/openwith/redirect/RedirectFixer.java
@@ -1,4 +1,4 @@
-package com.tasomaniac.openwith.resolver;
+package com.tasomaniac.openwith.redirect;
 
 import android.content.Intent;
 import android.net.Uri;
